+++ libraries/chain/include/node/chain/block_log.hpp
@@ -45,12 +45,12 @@ namespace node { namespace chain {
          uint64_t append( const signed_block& b );
          void flush();
          std::pair< signed_block, uint64_t > read_block( uint64_t file_pos )const;
-         optional< signed_block > read_block_by_num( uint32_t block_num )const;
+         optional< signed_block > read_block_by_num( uint64_t block_num )const;
 
          /**
           * Return offset of block in file, or block_log::npos if it does not exist.
           */
-         uint64_t get_block_pos( uint32_t block_num ) const;
+         uint64_t get_block_pos( uint64_t block_num ) const;
          signed_block read_head()const;
          const optional< signed_block >& head()const;
